requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langgraph==0.0.62
langchain==0.1.20
langchain-google-genai==1.0.8
google-generativeai==0.3.2
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2 